Cleaning up db context initialisation for SqLite Adding pre test run checks for Cosmos

+++ Nethereum.BlockchainStore.CosmosCore.Tests/CosmosCoreRepositoryTests.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using Nethereum.BlockchainStore.CosmosCore.Bootstrap;
 using Nethereum.BlockchainStore.Test.Base.RepositoryTests;
@@ -12,18 +14,30 @@ namespace Nethereum.BlockchainStore.CosmosCore.Tests
     public class CosmosFixture : IDisposable
     {
         public static readonly string[] CommandLineArgs = new string[] {};
+        public const string CosmosEmulatorBaseUrl = "https://localhost:8081";
+        public const string CosmosAccessKey =
+            "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
+
+        public static Uri CosmosEmulatorUi = new Uri($"{CosmosEmulatorBaseUrl}/_explorer/index.html");
+        public const string EmulatorPath = @"C:\Program Files\Azure Cosmos DB Emulator\CosmosDB.Emulator.exe";
 
         public CosmosFixture()
         {
             ConfigurationUtils.SetEnvironment("development");
             var appConfig = ConfigurationUtils.Build(CommandLineArgs);
 
-            //Azure Cosmos DB emulator settings
-            //https://localhost:8081/_explorer/index.html
+            appConfig.SetCosmosEndpointUri(CosmosEmulatorBaseUrl);
+            appConfig.SetCosmosAccessKey(CosmosAccessKey);
 
-            appConfig.SetCosmosEndpointUri("https://localhost:8081");
-            appConfig.SetCosmosAccessKey(
-                "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==");
+            if (!IsEmulatorRunning())
+            {
+                string message = $"The Cosmos DB Emulator is not running at Uri '{CosmosEmulatorBaseUrl}'";
+
+                if(File.Exists(EmulatorPath))
+                    throw new Exception(message);
+
+                throw new Exception($"{message}.  The emulator does not appear to be installed (at least not in the default location '{EmulatorPath}').");
+            }
 
             Factory = CosmosRepositoryFactory.Create(appConfig, deleteAllExistingCollections: true);
         }
@@ -33,6 +47,24 @@ namespace Nethereum.BlockchainStore.CosmosCore.Tests
         public void Dispose()
         {
             Factory?.DeleteDatabase().Wait();
+        }
+
+        private bool IsEmulatorRunning()
+        {
+            try
+            {
+                using (var webClient = new WebClient())
+                {
+                    using (webClient.OpenRead(CosmosEmulatorUi))
+                    {
+                        return true;
+                    }
+                }
+            }
+            catch (WebException)
+            {
+                return false;
+            }
         }
     }
